# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Meeting Slot Service API
  version: 1.0.0
  description: |
    REST API for managing meeting slots and finding optimal meeting times for distributed teams.
    
    This service helps teams across multiple time zones find the best available meeting slots
    based on participant availability, working hours, and configurable preferences.
servers:
  - url: https://{api-gateway-id}.execute-api.{region}.amazonaws.com/v1
    description: AWS API Gateway
    variables:
      api-gateway-id:
        default: xxxxx
        description: API Gateway ID
      region:
        default: us-east-1
        description: AWS Region
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/events:
    post:
      tags:
        - Events
      summary: Create a new event
      description: Creates a new meeting event with proposed time slots
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
            example:
              title: "Weekly Team Standup"
              description: "Weekly synchronization meeting"
              organizer_id: "usr_abc123"
              duration_minutes: 60
              proposed_slots:
                - start_time: "2026-02-01T14:00:00+05:30"
                  end_time: "2026-02-01T16:00:00+05:30"
                  timezone: "Asia/Kolkata"
                - start_time: "2026-02-01T17:00:00+05:30"
                  end_time: "2026-02-01T19:00:00+05:30"
                  timezone: "Asia/Kolkata"
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
              example:
                success: true
                data:
                  id: "evt_xyz789"
                  title: "Weekly Team Standup"
                  description: "Weekly synchronization meeting"
                  organizer_id: "usr_abc123"
                  duration_minutes: 60
                  status: "pending"
                  proposed_slots:
                    - id: 1
                      event_id: "evt_xyz789"
                      start_time: "2026-02-01T14:00:00+05:30"
                      end_time: "2026-02-01T16:00:00+05:30"
                      timezone: "Asia/Kolkata"
                      created_at: "2026-02-18T10:30:00Z"
                    - id: 2
                      event_id: "evt_xyz789"
                      start_time: "2026-02-01T17:00:00+05:30"
                      end_time: "2026-02-01T19:00:00+05:30"
                      timezone: "Asia/Kolkata"
                      created_at: "2026-02-18T10:30:00Z"
                  participants: []
                  created_at: "2026-02-18T10:30:00Z"
                  updated_at: "2026-02-18T10:30:00Z"
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Events
      summary: List events
      description: Retrieves a paginated list of events with optional filters
      operationId: listEvents
      parameters:
        - name: organizer_id
          in: query
          description: Filter by organizer ID
          schema:
            type: string
          example: "usr_abc123"
        - name: status
          in: query
          description: Filter by event status
          schema:
            type: string
            enum: [pending, confirmed, cancelled]
          example: "pending"
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
              example:
                success: true
                data:
                  - id: "evt_xyz789"
                    title: "Weekly Team Standup"
                    description: "Weekly synchronization meeting"
                    organizer_id: "usr_abc123"
                    duration_minutes: 60
                    status: "pending"
                    proposed_slots: []
                    participants: []
                    created_at: "2026-02-18T10:30:00Z"
                    updated_at: "2026-02-18T10:30:00Z"
                pagination:
                  page: 1
                  limit: 20
                  total: 1
  /api/v1/events/{id}:
    get:
      tags:
        - Events
      summary: Get event by ID
      description: Retrieves a specific event by its unique identifier
      operationId: getEventById
      parameters:
        - $ref: '#/components/parameters/EventIdParam'
      responses:
        '200':
          description: Event found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
              example:
                success: true
                data:
                  id: "evt_xyz789"
                  title: "Weekly Team Standup"
                  description: "Weekly synchronization meeting"
                  organizer_id: "usr_abc123"
                  duration_minutes: 60
                  status: "pending"
                  proposed_slots:
                    - id: 1
                      event_id: "evt_xyz789"
                      start_time: "2026-02-01T14:00:00+05:30"
                      end_time: "2026-02-01T16:00:00+05:30"
                      timezone: "Asia/Kolkata"
                      created_at: "2026-02-18T10:30:00Z"
                  participants:
                    - id: 1
                      event_id: "evt_xyz789"
                      user_id: "usr_def456"
                      status: "invited"
                      created_at: "2026-02-18T10:35:00Z"
                      updated_at: "2026-02-18T10:35:00Z"
                      user:
                        id: "usr_def456"
                        name: "Jane Smith"
                        email: "jane.smith@example.com"
                        created_at: "2026-02-18T11:00:00Z"
                        updated_at: "2026-02-18T11:00:00Z"
                  created_at: "2026-02-18T10:30:00Z"
                  updated_at: "2026-02-18T10:30:00Z"
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "NOT_FOUND"
                  message: "Event not found"

    put:
      tags:
        - Events
      summary: Update event
      description: Updates an existing event's information
      operationId: updateEvent
      parameters:
        - $ref: '#/components/parameters/EventIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventRequest'
            example:
              title: "Weekly Team Sync - Updated"
              description: "Updated meeting description"
              duration_minutes: 45
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Events
      summary: Delete event
      description: Soft deletes an event by its unique identifier
      operationId: deleteEvent
      parameters:
        - $ref: '#/components/parameters/EventIdParam'
      responses:
        '204':
          description: Event deleted successfully
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/events/{id}/participants:
    post:
      tags:
        - Participants
      summary: Add participants to event
      description: Adds one or more participants to an event
      operationId: addParticipants
      parameters:
        - $ref: '#/components/parameters/EventIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddParticipantRequest'
            example:
              user_ids: ["usr_def456", "usr_ghi789", "usr_jkl012"]
      responses:
        '201':
          description: Participants added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddParticipantResponse'
              example:
                success: true
                data:
                  message: "Participants processed"
                  added_count: 3
                  failed_count: 0
                  added_user_ids: ["usr_def456", "usr_ghi789", "usr_jkl012"]
                  failed_user_ids: []
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Participants
      summary: Get event participants
      description: Retrieves all participants for a specific event
      operationId: getEventParticipants
      parameters:
        - $ref: '#/components/parameters/EventIdParam'
      responses:
        '200':
          description: List of participants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticipantListResponse'
              example:
                success: true
                data:
                  - id: 1
                    event_id: "evt_xyz789"
                    user_id: "usr_def456"
                    status: "invited"
                    created_at: "2026-02-18T10:35:00Z"
                    updated_at: "2026-02-18T10:35:00Z"
                    user:
                      id: "usr_def456"
                      name: "Jane Smith"
                      email: "jane.smith@example.com"
                      created_at: "2026-02-18T11:00:00Z"
                      updated_at: "2026-02-18T11:00:00Z"
                  - id: 2
                    event_id: "evt_xyz789"
                    user_id: "usr_ghi789"
                    status: "accepted"
                    created_at: "2026-02-18T10:35:00Z"
                    updated_at: "2026-02-18T12:00:00Z"
                    user:
                      id: "usr_ghi789"
                      name: "Bob Johnson"
                      email: "bob.j@example.com"
                      created_at: "2026-02-18T11:15:00Z"
                      updated_at: "2026-02-18T11:15:00Z"
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/events/{id}/participants/{user_id}:
    delete:
      tags:
        - Participants
      summary: Remove participant from event
      description: Removes a participant from an event
      operationId: removeParticipant
      parameters:
        - $ref: '#/components/parameters/EventIdParam'
        - $ref: '#/components/parameters/UserIdPathParam'
      responses:
        '204':
          description: Participant removed successfully
        '404':
          description: Participant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
...
    UserIdPathParam:
      name: user_id
      in: path
      required: true
      description: User unique identifier
      schema:
        type: string
        example: "usr_def456"
    EventIdParam:
      name: id
      in: path
      required: true
      description: Event unique identifier
      schema:
        type: string
        example: "evt_xyz789"
    PageParam:
      name: page
      in: query
      description: Page number (default 1)
      schema:
        type: integer
        default: 1
        minimum: 1
      example: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page (default 20, max 100)
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      example: 20
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/Error'
    Error:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: "NOT_FOUND"
        message:
          type: string
          description: Error message
          example: "Resource not found"
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
    TimeSlot:
      type: object
      required:
        - start_time
        - end_time
        - timezone
      properties:
        start_time:
          type: string
          format: date-time
          description: Slot start time
          example: "2026-02-20T09:00:00Z"
        end_time:
          type: string
          format: date-time
          description: Slot end time
          example: "2026-02-20T10:00:00Z"
        timezone:
          type: string
          description: Timezone
          example: "UTC"

    # User Schemas
    User:
      type: object
      properties:
        id:
          type: string
          description: Unique user identifier
          example: "usr_abc123"
        name:
          type: string
          description: User full name
          example: "John Doe"
        email:
          type: string
          format: email
          description: User email address
          example: "john.doe@example.com"
        created_at:
          type: string
          format: date-time
          description: User creation timestamp
          example: "2026-02-18T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: User last update timestamp
          example: "2026-02-18T10:30:00Z"
...
    Event:
      type: object
      properties:
        id:
          type: string
          description: Unique event identifier
          example: "evt_xyz789"
        title:
          type: string
          description: Event title
          example: "Weekly Team Standup"
        description:
          type: string
          description: Event description
          example: "Weekly synchronization meeting"
        organizer_id:
          type: string
          description: ID of the event organizer
          example: "usr_abc123"
        duration_minutes:
          type: integer
          description: Event duration in minutes
          example: 60
        status:
          type: string
          enum: [pending, confirmed, cancelled]
          description: Event status
          example: "pending"
        proposed_slots:
          type: array
          items:
            $ref: '#/components/schemas/ProposedSlot'
        participants:
          type: array
          items:
            $ref: '#/components/schemas/Participant'
        created_at:
          type: string
          format: date-time
          example: "2026-02-18T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-02-18T10:30:00Z"
    ProposedSlot:
      type: object
      properties:
        id:
          type: integer
          description: Slot ID
          example: 1
        event_id:
          type: string
          description: Associated event ID
          example: "evt_xyz789"
        start_time:
          type: string
          format: date-time
          description: Slot start time
          example: "2026-02-01T14:00:00+05:30"
        end_time:
          type: string
          format: date-time
          description: Slot end time
          example: "2026-02-01T16:00:00+05:30"
        timezone:
          type: string
          description: Timezone
          example: "Asia/Kolkata"
        created_at:
          type: string
          format: date-time
          example: "2026-02-18T10:30:00Z"
    CreateEventRequest:
      type: object
      required:
        - title
        - organizer_id
        - duration_minutes
      properties:
        title:
          type: string
          description: Event title
          example: "Weekly Team Standup"
        description:
          type: string
          description: Event description
          example: "Weekly synchronization meeting"
        organizer_id:
          type: string
          description: ID of the event organizer
          example: "usr_abc123"
        duration_minutes:
          type: integer
          minimum: 5
          maximum: 480
          description: Event duration in minutes
          example: 60
        proposed_slots:
          type: array
          items:
            $ref: '#/components/schemas/TimeSlot'
    UpdateEventRequest:
      type: object
      properties:
        title:
          type: string
          description: Event title
          example: "Weekly Team Sync - Updated"
        description:
          type: string
          description: Event description
          example: "Updated meeting description"
        duration_minutes:
          type: integer
          minimum: 5
          maximum: 480
          description: Event duration in minutes
          example: 45
    EventResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Event'
    EventListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Participant Schemas
    Participant:
      type: object
      properties:
        id:
          type: integer
          description: Unique participant record identifier
          example: 1
        event_id:
          type: string
          description: Associated event ID
          example: "evt_xyz789"
        user_id:
          type: string
          description: Associated user ID
          example: "usr_def456"
        status:
          type: string
          enum: [invited, accepted, declined, responded]
          description: Participant's response status
          example: "invited"
        user:
          $ref: '#/components/schemas/User'
        created_at:
          type: string
          format: date-time
          example: "2026-02-18T10:35:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-02-18T10:35:00Z"
    AddParticipantRequest:
      type: object
      required:
        - user_ids
      properties:
        user_ids:
          type: array
          description: Array of User IDs to add as participants
          items:
            type: string
          example: ["usr_def456", "usr_ghi789"]
    AddParticipantResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            message:
              type: string
              example: "Participants processed"
            added_count:
              type: integer
              example: 2
            failed_count:
              type: integer
              example: 0
            added_user_ids:
              type: array
              items:
                type: string
              example: ["usr_def456", "usr_ghi789"]
            failed_user_ids:
              type: array
              items:
                type: object
                properties:
                  user_id:
                    type: string
                  error:
                    type: string
              example: []
    ParticipantListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Participant'

    # Availability Schemas
...